refactor: merge tinygo files
